% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf.R
\name{matrix_maximise}
\alias{matrix_maximise}
\title{Transform matrix according to maximise parameter}
\usage{
matrix_maximise(x, maximise)
}
\arguments{
\item{x}{(\code{data.frame()}|\code{matrix()}) A numerical matrix or \code{data.frame}.}

\item{maximise}{(\code{logical()} | \code{logical(1)}) \cr Whether the objectives must be
maximised instead of minimised. Either a single logical value that applies
to all objectives or a vector of logical values, with one value per
objective.}
}
\value{
\code{x} transformed such that every column where \code{maximise} is \code{TRUE} is multiplied by \code{-1}.
}
\description{
Transform matrix according to maximise parameter
}
